move daily weather into daily subcommand and use constant for prototype
