+++ src/main/java/com/gorlah/kappabot/KappaBotApplication.java
@@ -1,5 +1,6 @@
 package com.gorlah.kappabot;
 
+import org.springframework.beans.factory.config.ConfigurableBeanFactory;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.annotation.Bean;
@@ -14,7 +15,7 @@ public class KappaBotApplication {
     }
 
     @Bean
-    @Scope("prototype")
+    @Scope(ConfigurableBeanFactory.SCOPE_PROTOTYPE)
     RestTemplate restTemplate() {
         return new RestTemplate();
     }
