+++ src/main/java/com/gorlah/kappabot/subcommand/WeatherCommand.java
@@ -1,13 +1,8 @@
 package com.gorlah.kappabot.subcommand;
 
-import com.google.common.base.Joiner;
-import com.gorlah.kappabot.command.Command;
-import com.gorlah.kappabot.integration.openweather.OpenWeatherFetcher;
 import com.gorlah.kappabot.integration.openweather.OpenWeatherIntegration;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
-
-import java.util.ArrayList;
 
 
 @Component
@@ -15,7 +10,6 @@ import java.util.ArrayList;
 public class WeatherCommand extends Subcommand {
 
     private final OpenWeatherIntegration openWeatherIntegration;
-    private final OpenWeatherFetcher openWeatherFetcher;
 
     @Override
     public String getName() {
@@ -24,7 +18,7 @@ public class WeatherCommand extends Subcommand {
 
     @Override
     public String getHelpText() {
-        return "Gets the weather by ZIP Code™ or city name.";
+        return "Gets information about the weather.";
     }
 
     @Override
@@ -40,14 +34,5 @@ public class WeatherCommand extends Subcommand {
     @Override
     public boolean isEnabled() {
         return openWeatherIntegration.isEnabled();
-    }
-
-    @Override
-    public String process(Command command, ArrayList<String> parameters) {
-        var zipCodeOrCityName = Joiner.on(' ').join(parameters);
-        if (zipCodeOrCityName.isBlank()) {
-            return "Must specify a ZIP Code™ or city name!";
-        }
-        return openWeatherFetcher.fetchWeather(zipCodeOrCityName);
     }
 }
